requirements.txt
langchain-community==0.0.14
langgraph==0.0.15
crewAI==0.11.0
google-search-results==2.1.0
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
tavily-python==0.3.1
beautifulsoup4==4.12.3
python-dotenv==1.0.0